// src/client/src/app/modules/admin/catalog/models/variantReport.ts
export interface VaraintReport {
    availableQuantity: number;
    barcodeSymbology: string;
    discountFactor: number;
    lastUpdatedOn: string;
    name: string;
    productId: string;
    productCode: string;
    promotionMode: number;
    combinePricePkr: number;
    afterDiscountPrices: number;
}